Split VaultManager into focused extensions
- Extract VaultManager+KeySharing.swift (186 lines): GRDB observation, key sharing, import logic
- Extract VaultManager+Pairing.swift (209 lines): initiator + joiner pairing flows
- VaultManager.swift reduced from 805 to 432 lines (coordinator, lifecycle, device management)
- Internal helper methods use package-internal access (no public API change)
- 390 tests passing, 0 lint violations
