+++ ConvosCore/Sources/ConvosCore/Vault/VaultManager+KeySharing.swift
@@ -1,0 +1,186 @@
+import Combine
+import Foundation
+import GRDB
+@preconcurrency import XMTPiOS
+
+extension VaultManager {
+    func startObservingInboxes() {
+        let observation = ValueObservation.tracking { db in
+            try Self.fetchUnsharedRows(db)
+        }
+
+        inboxObservationCancellable = observation
+            .publisher(in: databaseReader)
+            .sink(
+                receiveCompletion: { _ in },
+                receiveValue: { [weak self] unsharedRows in
+                    guard let self, !unsharedRows.isEmpty else { return }
+                    Task { await self.shareUnsharedInboxes(unsharedRows) }
+                }
+            )
+        Log.info("Vault: started observing inboxes for key sharing")
+    }
+
+    func shareUnsharedInboxes(_ rows: [InboxConversationRow]) async {
+        guard await isConnected, hasMultipleDevices, let databaseWriter else { return }
+        Log.info("Vault: found \(rows.count) unshared inbox(es) with conversations")
+
+        for row in rows {
+            await shareKeyForInbox(row.inboxId, clientId: row.clientId)
+
+            try? await databaseWriter.write { db in
+                try db.execute(
+                    sql: "UPDATE inbox SET sharedToVault = 1 WHERE inboxId = ?",
+                    arguments: [row.inboxId]
+                )
+            }
+        }
+    }
+
+    func shareKeyForInbox(_ inboxId: String, clientId: String) async {
+        guard let identity = try? await identityStore.identity(for: inboxId) else {
+            Log.warning("Vault: no keychain identity for inbox \(inboxId)")
+            return
+        }
+
+        let conversationId = try? await databaseReader.read { db -> String? in
+            let sql = """
+                SELECT c.id FROM conversation c
+                WHERE c.clientId = ? AND c.id NOT LIKE 'draft-%'
+                LIMIT 1
+                """
+            return try String.fetchOne(db, sql: sql, arguments: [clientId])
+        }
+
+        let keyInfo = InboxKeyInfo(
+            inboxId: inboxId,
+            clientId: clientId,
+            conversationId: conversationId ?? "",
+            privateKeyData: Data(identity.keys.privateKey.secp256K1.bytes),
+            databaseKey: identity.keys.databaseKey
+        )
+        Log.info("Vault: sharing key for inbox \(inboxId)")
+        await shareKeyFromNotification(keyInfo)
+    }
+
+    func shareKeyFromNotification(_ keyInfo: InboxKeyInfo) async {
+        guard await isConnected, hasMultipleDevices, let installationId = await vaultClient.installationId else { return }
+        let share = DeviceKeyShareContent(
+            conversationId: keyInfo.conversationId,
+            inboxId: keyInfo.inboxId,
+            clientId: keyInfo.clientId,
+            privateKeyData: keyInfo.privateKeyData,
+            databaseKey: keyInfo.databaseKey,
+            senderInstallationId: installationId,
+            senderDeviceName: deviceName
+        )
+
+        do {
+            try await vaultClient.send(share, codec: DeviceKeyShareCodec())
+        } catch {
+            Log.error("Vault: failed to share key for \(keyInfo.inboxId): \(error)")
+        }
+    }
+
+    func checkUnsharedInboxes() async {
+        guard await isConnected, hasMultipleDevices else { return }
+        guard let rows = try? await databaseReader.read({ db in
+            try Self.fetchUnsharedRows(db)
+        }), !rows.isEmpty else { return }
+        await shareUnsharedInboxes(rows)
+    }
+
+    static func fetchUnsharedRows(_ db: Database) throws -> [InboxConversationRow] {
+        let sql = """
+            SELECT i.inboxId, i.clientId, c.id as conversationId
+            FROM inbox i
+            INNER JOIN conversation c ON c.clientId = i.clientId
+                AND c.id NOT LIKE 'draft-%'
+                AND c.isUnused = 0
+            WHERE i.isVault = 0 AND i.sharedToVault = 0
+            """
+        return try Row.fetchAll(db, sql: sql).map { row in
+            InboxConversationRow(
+                inboxId: row["inboxId"],
+                clientId: row["clientId"],
+                conversationId: row["conversationId"] ?? ""
+            )
+        }
+    }
+
+    func importKeyShare(_ share: DeviceKeyShareContent) async {
+        guard await !hasIdentity(for: share.inboxId) else { return }
+
+        do {
+            let keys = try KeychainIdentityKeys(
+                privateKeyData: share.privateKeyData,
+                databaseKey: share.databaseKey
+            )
+            let identity = try await identityStore.save(
+                inboxId: share.inboxId,
+                clientId: share.clientId,
+                keys: keys
+            )
+
+            if let databaseWriter {
+                let inboxWriter = InboxWriter(dbWriter: databaseWriter)
+                try await inboxWriter.save(inboxId: identity.inboxId, clientId: identity.clientId)
+            }
+
+            postImportNotification(inboxId: identity.inboxId, clientId: identity.clientId)
+        } catch {
+            Log.error("Vault: failed to import key share for \(share.inboxId): \(error)")
+        }
+    }
+
+    func importKeyBundle(_ bundle: DeviceKeyBundleContent) async {
+        var importedEntries: [(inboxId: String, clientId: String)] = []
+        for key in bundle.keys {
+            guard await !hasIdentity(for: key.inboxId) else { continue }
+
+            do {
+                let keys = try KeychainIdentityKeys(
+                    privateKeyData: key.privateKeyData,
+                    databaseKey: key.databaseKey
+                )
+                let identity = try await identityStore.save(
+                    inboxId: key.inboxId,
+                    clientId: key.clientId,
+                    keys: keys
+                )
+
+                if let databaseWriter {
+                    let inboxWriter = InboxWriter(dbWriter: databaseWriter)
+                    try await inboxWriter.save(inboxId: identity.inboxId, clientId: identity.clientId)
+                }
+
+                importedEntries.append((inboxId: identity.inboxId, clientId: identity.clientId))
+            } catch {
+                Log.error("Vault: failed to import key bundle entry for \(key.inboxId): \(error)")
+            }
+        }
+
+        if !importedEntries.isEmpty {
+            for entry in importedEntries {
+                postImportNotification(inboxId: entry.inboxId, clientId: entry.clientId)
+            }
+            NotificationCenter.default.post(
+                name: .vaultDidReceiveKeyBundle,
+                object: nil,
+                userInfo: ["importedCount": importedEntries.count]
+            )
+        }
+    }
+
+    func postImportNotification(inboxId: String, clientId: String) {
+        NotificationCenter.default.post(
+            name: .vaultDidImportInbox,
+            object: nil,
+            userInfo: ["inboxId": inboxId, "clientId": clientId]
+        )
+    }
+
+    func hasIdentity(for inboxId: String) async -> Bool {
+        (try? await identityStore.identity(for: inboxId)) != nil
+    }
+}
